Use pthread mutexes on linux
kfreebsd/hurd lack robust pthred mutexes. Therefore we need to keep
using fcntl there. For fcnl, we need to specify the directory in
apache2.conf explicitly, to support multi-instance setups. To achieve
this, we add a marker to apache2.conf that is removed or replaced by
a '#' during build, depending on the OS.