Fix conffiles from wheezy upgrade
Some conffile may have got the wrong content during upgrade from
wheezy to early jessie versions. In order to fix this without dpkg
prompting the user, we need to replace those conffiles during preinst.

Unfortunately, in preinst the package is not yet unpacked so we cannot
access any of its files. Therefore, the correct conffile content is
embedded in a base64 encoded tarball in the preinst script.