apache2_switch_mpm: check if admin has disabled the mpm
Don't override the local admin's action if the requested mpm has been
disabled by the admin.

This may not be perfect, because we don't check if the current mpm has
been enabled by the admin. But on the other hand, the admin-installed
state is also set for the mpm that has been active during the 2.2 -> 2.4
update. So there may not be an alternative.