Restart htcacheclean on upgrade
postinst's call to start_htcacheclean() missed the $@ arguments, so it
would always start not restart (and the former does nothing if
htcacheclean is already running).

The check for cache_disk.load should not be necessary because
invoke-rc.d should do the correct thing(tm).

Also add a comment to prerm.

Avoid that apache2-utils is upgraded without htcacheclean being
restarted by tightening the dependency from apache2 to apache2-utils.