Fix test suite to actually run
It turns out the test suite refuses to run as root but returns true even
in that case. Fix this by creating a temp user.

It seems this has been broken since 2.4.27-4, where the test suite had
been updated and the the debci test duration dropped from 15min to 3min.
Also, don't rely on the exit status anymore but parse the test output.